backend/requirements.txt
fastapi
uvicorn
gitpython
radon
transformers
torch
networkx
scikit-learn
matplotlib
reportlab
pydot
